+++ R/sharp-package.R
@@ -10,8 +10,8 @@
 #' medoids, K means or Gaussian mixture models in consensus clustering.
 #'
 #' \tabular{ll}{ Package: \tab sharp\cr Type: \tab Package\cr Version: \tab
-#' 1.4.0 \cr Date: \tab 2023-04-18 \cr License: \tab GPL (>= 3)\cr Maintainer:
-#' \tab Barbara Bodinier \email{b.bodinier@@imperial.ac.uk}}
+#' 1.4.1 \cr Date: \tab 2023-05-30 \cr License: \tab GPL (>= 3)\cr Maintainer:
+#' \tab Barbara Bodinier \email{b.bodinier@@gmail.com}}
 #'
 #' @references \insertRef{OurConsensusClustering}{sharp}
 #'